Secure websocket transactions

+++ app/models/Todo.scala
@@ -21,6 +21,8 @@ case class Todo(id:String,text:String,done:Boolean,disp_order:Int,userId:String)
 
 object Todo {
 
+   var userId:String=null
+
   /* Parse a todo from the result set */
   def simple={
     get[String]("id")~
@@ -34,6 +36,7 @@ object Todo {
 
   /* Retrieve all todos */
   def getAll(userId:String):Seq[Todo]={
+    this.userId=userId
     DB.withConnection{implicit connection=>
       SQL("SELECT * FROM Todo WHERE userId={userId}").on('userId->userId).as(Todo.simple *)
     }
@@ -50,40 +53,54 @@ object Todo {
 
   /* Add a todo to the database */
   def create(task:Todo)={
-    DB.withConnection{implicit connection=>
-      SQL("INSERT INTO Todo (id,text,done,disp_order,userId) VALUES({id},{text},{done},{disp_order},{userId})").on(
-        'id->task.id,
-        'text->task.text,
-        'done->task.done,
-        'disp_order->task.disp_order,
-        'userId->task.userId
-      ).executeUpdate()
+    if (this.userId==task.userId){
+      DB.withConnection{implicit connection=>
+        SQL("INSERT INTO Todo (id,text,done,disp_order,userId) VALUES({id},{text},{done},{disp_order},{userId})").on(
+          'id->task.id,
+          'text->task.text,
+          'done->task.done,
+          'disp_order->task.disp_order,
+          'userId->task.userId
+        ).executeUpdate()
+      }
     }
   }
 
   /* Delete a todo by its id */
-  def delete(id:String)={
-    DB.withConnection{implicit connection=>
-      SQL("DELETE FROM Todo WHERE id={id}").on(
-        'id->id).executeUpdate()
+  def delete(id:String,userId:String)={
+    if (this.userId==userId){
+      DB.withConnection{implicit connection=>
+        SQL("DELETE FROM Todo WHERE id={id} AND userId={userId}").on(
+          'id->id,
+          'userId->userId
+        ).executeUpdate()
+      }
     }
   }
 
   /* Update a todo text */
-  def updateText(id:String,text:String)={
-    DB.withConnection{implicit connection=>
-      SQL("UPDATE Todo SET text={text} WHERE id={id}").on(
-        'id->id,'text->text
-      ).executeUpdate()
+  def updateText(id:String,text:String,userId:String)={
+    if (this.userId==userId){
+      DB.withConnection{implicit connection=>
+        SQL("UPDATE Todo SET text={text} WHERE id={id} AND userId={userId}").on(
+          'id->id,
+          'text->text,
+          'userId->userId
+        ).executeUpdate()
+      }
     }
   }
 
   /* Update a todo done attribute */
-  def updateStatus(id:String,done:Boolean)={
-    DB.withConnection{implicit connection=>
-      SQL("UPDATE Todo SET done={done} WHERE id={id}").on(
-        'id->id,'done->done
-      ).executeUpdate()
+  def updateStatus(id:String,done:Boolean,userId:String)={
+    if (this.userId==userId){
+      DB.withConnection{implicit connection=>
+        SQL("UPDATE Todo SET done={done} WHERE id={id} AND userId={userId}").on(
+          'id->id,
+          'done->done,
+          'userId->userId
+        ).executeUpdate()
+      }
     }
   }
